+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/services/TransactionStatusTest.java
@@ -67,9 +67,9 @@ package com.radixdlt.integration.services;
 import com.google.inject.Scopes;
 import com.google.inject.multibindings.Multibinder;
 import com.radixdlt.api.data.TransactionStatus;
-import com.radixdlt.api.store.berkeley.BerkeleyTransactionsByIdStore;
-import com.radixdlt.api.transactions.TransactionStatusService;
-import com.radixdlt.api.transactions.TransactionStatusServiceModule;
+import com.radixdlt.api.transactions.lookup.BerkeleyTransactionsByIdStore;
+import com.radixdlt.api.transactions.lookup.TransactionStatusService;
+import com.radixdlt.api.transactions.lookup.TransactionStatusServiceModule;
 import com.radixdlt.engine.RadixEngine;
 import com.radixdlt.environment.Environment;
 import com.radixdlt.environment.deterministic.MultiNodeDeterministicRunner;
