+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/tokens/ArchiveTokensApiModule.java
@@ -1,10 +1,9 @@
-/* Copyright 2021 Radix Publishing Ltd incorporated in Jersey (Channel Islands).
- *
+/*
+ * Copyright 2021 Radix DLT Ltd incorporated in England.
  * Licensed under the Radix License, Version 1.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at:
  *
  * radixfoundation.org/licenses/LICENSE-v1
- *
  * The Licensor hereby grants permission for the Canonical version of the Work to be
  * published, distributed and used under or by reference to the Licensor’s trademark
  * Radix ® and use of any unregistered trade names, logos or get-up.
@@ -62,43 +61,46 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.handler;
+package com.radixdlt.api.tokens;
 
-import com.radixdlt.utils.functional.Result;
-import org.json.JSONObject;
-
-import com.google.inject.Inject;
-import com.google.inject.Singleton;
-import com.radixdlt.api.transactions.TransactionStatusService;
-import com.radixdlt.identifiers.AID;
+import com.google.inject.AbstractModule;
+import com.google.inject.Scopes;
+import com.google.inject.multibindings.Multibinder;
+import com.google.inject.multibindings.ProvidesIntoMap;
+import com.google.inject.multibindings.StringMapKey;
+import com.radixdlt.api.JsonRpcHandler;
+import com.radixdlt.api.qualifier.ArchiveEndpoint;
+import com.radixdlt.identifiers.REAddr;
+import com.radixdlt.networks.Addressing;
+import com.radixdlt.store.berkeley.BerkeleyAdditionalStore;
 
 import static com.radixdlt.api.JsonRpcUtil.withRequiredStringParameter;
-import static com.radixdlt.api.data.ApiErrors.UNKNOWN_TX_ID;
 
-@Singleton
-public class ArchiveTransactionsHandler {
-	private final TransactionStatusService transactionStatusService;
-
-	@Inject
-	public ArchiveTransactionsHandler(TransactionStatusService transactionStatusService) {
-		this.transactionStatusService = transactionStatusService;
+public final class ArchiveTokensApiModule extends AbstractModule {
+	@Override
+	protected void configure() {
+		bind(BerkeleyResourceInfoStore.class).in(Scopes.SINGLETON);
+		Multibinder.newSetBinder(binder(), BerkeleyAdditionalStore.class)
+			.addBinding().to(BerkeleyResourceInfoStore.class);
 	}
 
-	public JSONObject handleTransactionsGetTransactionStatus(JSONObject request) {
-		return withRequiredStringParameter(
-			request,
-			"txID",
-			idString -> AID.fromString(idString)
-				.map(txId -> transactionStatusService.getTransactionStatus(txId).asJson().put("txID", txId))
-		);
+	@ArchiveEndpoint
+	@ProvidesIntoMap
+	@StringMapKey("tokens.get_native_token")
+	public JsonRpcHandler tokensGetNativeToken(BerkeleyResourceInfoStore store) {
+		return req -> store.getResourceInfo(REAddr.ofNativeToken()).orElseThrow();
 	}
 
-	public JSONObject handleTransactionsLookupTransaction(JSONObject request) {
-		return withRequiredStringParameter(
-			request,
-			"txID",
-			idString -> AID.fromString(idString)
-				.flatMap(txId -> Result.fromOptional(UNKNOWN_TX_ID.with(txId), transactionStatusService.getTransaction(txId)))
+	@ArchiveEndpoint
+	@ProvidesIntoMap
+	@StringMapKey("tokens.get_info")
+	public JsonRpcHandler tokensGetInfo(Addressing addressing, BerkeleyResourceInfoStore store) {
+		return req -> withRequiredStringParameter(
+			req,
+			"rri",
+			rri ->
+				addressing.forResources().parseFunctional(rri)
+					.map(e -> e.map((symbol, addr) -> store.getResourceInfo(addr).orElseThrow()))
 		);
 	}
 }
