+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/transactions/lookup/BerkeleyTransactionsByIdStore.java
@@ -1,11 +1,9 @@
 /*
- * Copyright 2021 Radix Publishing Ltd incorporated in Jersey (Channel Islands).
- *
+ * Copyright 2021 Radix DLT Ltd incorporated in England.
  * Licensed under the Radix License, Version 1.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at:
  *
  * radixfoundation.org/licenses/LICENSE-v1
- *
  * The Licensor hereby grants permission for the Canonical version of the Work to be
  * published, distributed and used under or by reference to the Licensor’s trademark
  * Radix ® and use of any unregistered trade names, logos or get-up.
@@ -63,7 +61,7 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.store.berkeley;
+package com.radixdlt.api.transactions.lookup;
 
 import com.google.inject.Inject;
 import com.google.inject.Provider;
