+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/transactions/lookup/TransactionStatusServiceModule.java
@@ -1,10 +1,9 @@
-/* Copyright 2021 Radix DLT Ltd incorporated in England.
- *
+/*
+ * Copyright 2021 Radix DLT Ltd incorporated in England.
  * Licensed under the Radix License, Version 1.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at:
  *
  * radixfoundation.org/licenses/LICENSE-v1
- *
  * The Licensor hereby grants permission for the Canonical version of the Work to be
  * published, distributed and used under or by reference to the Licensor’s trademark
  * Radix ® and use of any unregistered trade names, logos or get-up.
@@ -62,77 +61,37 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.module;
+package com.radixdlt.api.transactions.lookup;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
-import com.google.inject.multibindings.Multibinder;
-import com.google.inject.multibindings.ProvidesIntoMap;
-import com.google.inject.multibindings.StringMapKey;
-import com.radixdlt.api.JsonRpcHandler;
-import com.radixdlt.api.qualifier.DeveloperEndpoint;
-import com.radixdlt.api.store.berkeley.BerkeleyTransactionIndexStore;
-import com.radixdlt.store.berkeley.BerkeleyAdditionalStore;
-import com.radixdlt.utils.functional.Failure;
-import com.radixdlt.utils.functional.Result;
-import org.json.JSONArray;
+import com.google.inject.multibindings.ProvidesIntoSet;
+import com.radixdlt.environment.EventProcessorOnRunner;
+import com.radixdlt.environment.Runners;
+import com.radixdlt.mempool.MempoolAddFailure;
+import com.radixdlt.mempool.MempoolAddSuccess;
 
-import java.util.List;
-
-import static com.radixdlt.api.JsonRpcUtil.jsonObject;
-import static com.radixdlt.api.JsonRpcUtil.withRequiredParameters;
-
-public final class TransactionIndexApiModule extends AbstractModule {
+public class TransactionStatusServiceModule extends AbstractModule {
 	@Override
 	public void configure() {
-		bind(BerkeleyTransactionIndexStore.class).in(Scopes.SINGLETON);
-		Multibinder.newSetBinder(binder(), BerkeleyAdditionalStore.class)
-			.addBinding().to(BerkeleyTransactionIndexStore.class);
+		bind(TransactionStatusService.class).in(Scopes.SINGLETON);
 	}
 
-	@DeveloperEndpoint
-	@ProvidesIntoMap
-	@StringMapKey("index.get_transaction_count")
-	public JsonRpcHandler indexGetTransactionCount(BerkeleyTransactionIndexStore store) {
-		return request -> withRequiredParameters(
-			request,
-			List.of(),
-			params -> Result.wrap(
-				e -> Failure.failure(-1, e.getMessage()),
-				() -> {
-					var totalCount = store.getCount();
-					return jsonObject()
-						.put("totalCount", totalCount);
-				}
-			)
+	@ProvidesIntoSet
+	public EventProcessorOnRunner<?> mempoolAddFailureEventProcessor(TransactionStatusService transactionStatusService) {
+		return new EventProcessorOnRunner<>(
+			Runners.APPLICATION,
+			MempoolAddFailure.class,
+			transactionStatusService.mempoolAddFailureEventProcessor()
 		);
 	}
 
-	@DeveloperEndpoint
-	@ProvidesIntoMap
-	@StringMapKey("index.get_transactions")
-	public JsonRpcHandler indexGetTransactions(BerkeleyTransactionIndexStore store) {
-		return request -> withRequiredParameters(
-			request,
-			List.of("offset", "limit"),
-			params -> Result.wrap(
-				e -> Failure.failure(-1, e.getMessage()),
-				() -> {
-					var offset = params.getLong("offset");
-					var limit = params.getLong("limit");
-					var transactions = new JSONArray();
-					try (var stream = store.get(offset)) {
-						stream.limit(limit).forEach(transactions::put);
-					}
-					var totalCount = store.getCount();
-					var nextOffset = offset + transactions.length();
-					return jsonObject()
-						.put("transactions", transactions)
-						.put("count", transactions.length())
-						.put("totalCount", totalCount)
-						.put("nextOffset", nextOffset);
-				}
-			)
+	@ProvidesIntoSet
+	public EventProcessorOnRunner<?> mempoolAddSuccessEventProcessor(TransactionStatusService transactionStatusService) {
+		return new EventProcessorOnRunner<>(
+			Runners.APPLICATION,
+			MempoolAddSuccess.class,
+			transactionStatusService.mempoolAddSuccessEventProcessor()
 		);
 	}
 }
