+++ radixdlt-core/radixdlt/src/test/java/com/radixdlt/api/transactions/TransactionStatusServiceTest.java
@@ -73,7 +73,8 @@ import com.google.inject.multibindings.Multibinder;
 import com.radixdlt.SingleNodeAndPeersDeterministicNetworkModule;
 import com.radixdlt.api.chaos.mempoolfiller.MempoolFillerModule;
 import com.radixdlt.api.data.TransactionStatus;
-import com.radixdlt.api.store.berkeley.BerkeleyTransactionsByIdStore;
+import com.radixdlt.api.transactions.lookup.BerkeleyTransactionsByIdStore;
+import com.radixdlt.api.transactions.lookup.TransactionStatusService;
 import com.radixdlt.application.tokens.Amount;
 import com.radixdlt.atom.TxAction;
 import com.radixdlt.atom.TxnConstructionRequest;
